feat(select): add `multiple` option to limel-select, replacing limel-multi-select
BREAKING CHANGE: `limel-multi-select` has been removed.
Use `limel-select` with the `multiple` attribute instead.

fix #203
