--- src/components/multi-select/multi-select.mdx
@@ -1,13 +1,0 @@
----
-name: Multi Select
-route: /multi-select
-menu: Components
----
-
-# Multi Select
-
-<limel-props name="limel-multi-select" />
-
-## Example
-
-<limel-example name="limel-example-multi-select" />
