+++ src/components/select-multiple/select-multiple.scss
@@ -1,0 +1,26 @@
+@import '../../style/variables';
+
+@import '@lime-material-16px/form-field/mdc-form-field';
+@import '@lime-material-16px/checkbox/mdc-checkbox';
+@import "@lime-material-16px/floating-label/mdc-floating-label";
+
+.multi-select {
+    position: relative;
+}
+
+.multi-select-label {
+    padding-left: pxToRem(15);
+}
+
+.mdc-form-field {
+    display: flex;
+}
+
+.mdc-checkbox {
+    @include mdc-checkbox-ink-color(primary);
+    @include mdc-checkbox-container-colors(secondary, on-primary, secondary, on-primary);
+}
+
+limel-collapsible-section {
+    --body-max-height: 70vh;
+}
