+++ src/components/select-single/select-single.scss
