+++ src/components/select/select.mdx
@@ -14,6 +14,10 @@ When importing Option, see [Import Statements](/#import-statements).
 
 <limel-example name="limel-example-select" />
 
+### Select More Than One Option
+
+<limel-example name="limel-example-select-multiple" path="select" />
+
 ### Initially Empty
 
 <limel-example name="limel-example-select-initially-empty" path="select" />
