--- src/examples/multi-select/multi-select.scss
@@ -1,3 +1,0 @@
-p {
-    font-size: small;
-}
